# production_log_analyzer_backend/interfaces/openapi.runtime.json
{"openapi":"3.1.0","info":{"title":"Production Log Analyzer API","description":"Backend API for uploading production logs and generating a structured analysis report.\n\nNotes:\n- Evidence quotes are redacted to reduce risk of leaking sensitive data.\n- Root causes are presented as hypotheses unless directly evidenced in logs.\n- If timestamps are missing, the time window is reported as unknown.","version":"0.2.0"},"paths":{"/":{"get":{"tags":["Health"],"summary":"Health Check","description":"Health check endpoint.\n\nReturns:\n    JSON object indicating service health.","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check  Get"}}}}}}},"/api/logs/analyze":{"post":{"tags":["Log Analysis"],"summary":"Upload a log file (or zip of logs) and generate a structured analysis report","description":"Accepts a log file upload (text, JSON lines, or common log formats). Also accepts .zip archives containing log files (.log, .txt, .json, .ndjson). Returns a structured analysis report including summary statistics, issue clusters with evidence vs hypotheses, pattern detection, timeline buckets, and prioritized troubleshooting steps.","operationId":"upload_and_analyze_logs","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_and_analyze_logs"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadAnalyzeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_and_analyze_logs":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"Log file to analyze (text, JSON lines, or common log formats). May also be a .zip containing .log/.txt/.json/.ndjson files."}},"type":"object","required":["file"],"title":"Body_upload_and_analyze_logs"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UploadAnalyzeResponse":{"properties":{"filename":{"type":"string","title":"Filename","description":"Original uploaded filename."},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type","description":"Uploaded file content type if provided by client."},"parse_warnings":{"items":{"type":"string"},"type":"array","title":"Parse Warnings","description":"Any parsing warnings encountered."},"report":{"additionalProperties":true,"type":"object","title":"Report","description":"Structured analysis report following the production-log-analysis skill standards."}},"type":"object","required":["filename","report"],"title":"UploadAnalyzeResponse","description":"Structured analysis report response."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Health","description":"Service health and diagnostics."},{"name":"Log Analysis","description":"Upload logs and generate structured analysis reports."}]}